feat: separation of common functions
